webapi/tct-alarm-tizen-tests/alarm/AlarmRelative_constructor_delay_convert.html

<!DOCTYPE html>

<!--

Copyright (c) 2013 Samsung Electronics Co., Ltd.

Licensed under the Apache License, Version 2.0 (the License);
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.



Authors:
        Mariusz Polasinski <m.polasinski@samsung.com>


-->

<html>
<head>
<title>AlarmRelative_constructor_delay_convert</title>
<meta charset="utf-8">
<script type="text/javascript" src="../resources/testharness.js"></script>
<script type="text/javascript" src="../resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

//==== TEST: AlarmRelative_constructor_delay_convert
//==== LABEL Check conversion of delay parameter of AlarmRelative constructor
//==== SPEC Tizen Web API:Application:Alarm:AlarmRelative:constructor C
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/alarm.html
//==== PRIORITY P1
//==== TEST_CRITERIA CONSTRA CONSTRM

test(function () {
    var i, alarm1, alarm2, period = 300, delay = ["ABCD", null, undefined];
    for (i = 0; i < delay.length; i++) {
        alarm1 = new tizen.AlarmRelative(delay[i]);
        assert_true(alarm1 instanceof tizen.AlarmRelative, "object was not created properly");
        assert_equals(alarm1.delay, 0, "bad delay in the object");

        alarm2 = new tizen.AlarmRelative(delay[i], period);
        assert_true(alarm2 instanceof tizen.AlarmRelative, "object was not created properly");
        assert_equals(alarm2.delay, 0, "bad delay in the object");
    }
}, document.title);

</script>
</body>
</html>
